+++ src/Bbob/Cli/Remove.cs
@@ -6,9 +6,11 @@ public class Remove : Command
     public new static string Name => "remove";
     public new static string Help => "Remove the theme or plugin. Auto detect.\n" +
     "<option>:\n" +
-    "--global | -g : remove <content> from global directory.\n\n" +
+    "--all-theme : remove all theme.\n" +
+    "--all-plugin : remove all plugin.\n" +
+    "--global | -g : remove [name] or all plugin from global directory.\n\n" +
     "Use:\n" +
-    "// remove [option] <name>";
+    "// remove [option] [name]";
 
     private string _pluginOrTheme = string.Empty;
     public string PluginOrTheme
@@ -16,6 +18,10 @@ public class Remove : Command
         get => _pluginOrTheme;
         set => _pluginOrTheme = value.ToUpper();
     }
+
+    public bool isAllTheme { get; set; } = false;
+    public bool isAllPlugin { get; set; } = false;
+    private bool isAll => isAllTheme || isAllPlugin;
     private bool isGlobal = false;
     public bool Global
     {
@@ -41,10 +47,12 @@ public class Remove : Command
         public static string Themes = Path.Combine(Environment.CurrentDirectory, "themes");
         public readonly static string Temp = Path.Combine(Environment.CurrentDirectory, "temp");
     }
-    public Remove(string PluginOrTheme, bool Global) : base(false)
+    public Remove(string PluginOrTheme, bool Global, bool isAllPlugin, bool isAllTheme) : base(false)
     {
         this.PluginOrTheme = PluginOrTheme;
         this.Global = Global;
+        this.isAllPlugin = isAllPlugin;
+        this.isAllTheme = isAllTheme;
     }
 
     public override bool Process()
@@ -52,9 +60,41 @@ public class Remove : Command
         const string SUCCESS = "Success: ";
         const string FAILED = "Failed: ";
 
-        CliShared.TextType type = CliShared.isPluginOrThemeName(PluginOrTheme, out string fixedName);
-        PluginOrTheme = fixedName;
-        string directory = Path.Combine(type == CliShared.TextType.Plugin ? DownloadPath.Plugins : DownloadPath.Themes, PluginOrTheme);
+        CliShared.TextType type = isAllPlugin ? CliShared.TextType.Plugin : CliShared.TextType.Theme;
+        if (!isAll)
+        {
+            type = CliShared.isPluginOrThemeName(PluginOrTheme, out string fixedName);
+            PluginOrTheme = fixedName;
+        }
+        string root = type == CliShared.TextType.Theme ? DownloadPath.Themes : DownloadPath.Plugins;
+        string p = isGlobal ? "global" : "current";
+        string top = type == CliShared.TextType.Theme ? "theme" : "plugin";
+
+        if (isAll)
+        {
+            if (Directory.Exists(root))
+            {
+                if (type == CliShared.TextType.Plugin)
+                {
+                    Shared.SharedLib.DirectoryHelper.DeleteDirectory(root);
+                    Directory.CreateDirectory(root);
+                }
+                else
+                {
+                    var themes = ThemeProcessor.GetThemes();
+                    var filters = from theme in themes where theme.Info.name != "default" select theme;
+                    foreach (var filter in filters)
+                    {
+                        Shared.SharedLib.DirectoryHelper.DeleteDirectory(filter.Path);
+                    }
+                }
+            }
+
+            ConsoleHelper.printSuccess($"{SUCCESS}Remove all {top} from {p} directory.");
+            return true;
+        }
+
+        string directory = Path.Combine(root, PluginOrTheme);
         if (type == CliShared.TextType.None)
         {
             ConsoleHelper.printError($"{FAILED}Can't remove because it not plugin or theme.");
@@ -66,9 +106,7 @@ public class Remove : Command
             return false;
         }
         Shared.SharedLib.DirectoryHelper.DeleteDirectory(directory);
-        string p = isGlobal ? "global" : "current";
-        string top = type == CliShared.TextType.Theme ? "theme" : "plugin";
-        ConsoleHelper.printSuccess($"{SUCCESS}Remove {top} {PluginOrTheme} from {p} directory!");
+        ConsoleHelper.printSuccess($"{SUCCESS}Remove {top} {PluginOrTheme} from {p} directory.");
         return true;
     }
 }